Do not create empty clusters
If the k-means algorithm generates clusters that contain fewer points
than the number of degrees of freedom, then instead of creating empty
clusters, omit those clusters.

This ensures that the clustered KDE data structure is picklable and
eliminates a potential corner case that consumers of the clustered
KDE would otherwise have to deal with.