Robustly check for singular covariance matrix
The covariance matrix may be nonsingular even if the number of points
is greater than the number of degrees of freedom because the MCMC
chains contain many repeated samples, and small clusters may consist
only of repeated points.

This fixes a number of `np.linalg.LinAlgError` exceptions and positive
infinite BIC values that I found when I re-ran the "First Two Years"
events.