Don't create empty FITS header cards
Don't add `DISTMEAN` or `DISTSTD` cards to the FITS header if we are
not generating a 3D sky map. Don't add the OBJECT card if the user
did not pass a value for the `--objid` keyword argument.

If these are passed as `None` to `fits.write_sky_map`, then ugly empty
header cards are created. If we don't want to populate those header
cards, then we shouldn't pass them to begin with.

In full disclosure, this is due to an upstream change in how
`fits.write_sky_map` treats its keyword arguments.