Remove searched area calculation
We can make more accurate sky maps by using all of the posterior
samples for density estimation rather than holding half of them
in reserve for two-step greedy area calculation.

In the future, we should not calculate and quote two-step areas
and only use them for algorithm validation and testing, because
it is difficult to communicate, explain, reproduce, or use them
without including an auxilliary "search map," and moreover
because the differences between the FITS areas and the two-step
areas should be insignificant after the recent algorithm
improvements.