Restructure class hierarchy to add Python multiprocessing
Multiprocessing is used during the two most CPU-intensive steps:
the search over the number of clusters and extracting the distance
moments. This makes the code WAY faster on big, multi-core machines.
This will be important for rapidly generating results for low- or
medium-latency parameter estimation analyses.

Note that plotting the KDE assignments was removed as collateral
damage in making this change.