Make sure that both pickling and running with -j works
We have to put in some hooks to make instances of Clustered2DSkyKDE
picklable because we dynamically create subclasses with different
values of the `frame` class variable. This gets even trickier
because we need both the class and instance objects to be picklable.

I also added some more comments while I was fixing this.